Add json output so that other nb renderer plugins can present it as structured tree/table etc.
